+++ bitarray/interface.go
@@ -60,7 +60,7 @@ type BitArray interface {
 	// seen capacity of the sparse array.
 	Capacity() uint64
 	// Count returns the number of set bits in this array.
-	Count() uint64
+	Count() int
 	// Or will bitwise or the two bitarrays and return a new bitarray
 	// representing the result.
 	Or(other BitArray) BitArray
